Unify event types at aztec.js/api/events
