+++ yarn-project/aztec.js/src/wallet/wallet.ts
@@ -1,12 +1,11 @@
 import type { ChainInfo } from '@aztec/entrypoints/interfaces';
-import { BlockNumber, BlockNumberPositiveSchema } from '@aztec/foundation/branded-types';
 import type { Fr } from '@aztec/foundation/fields';
 import {
-  type AbiDecoded,
   AbiTypeSchema,
   type ContractArtifact,
   ContractArtifactSchema,
   type EventMetadataDefinition,
+  EventMetadataDefinitionSchema,
   type FunctionCall,
   FunctionType,
 } from '@aztec/stdlib/abi';
@@ -20,7 +19,7 @@ import {
   type ContractMetadata,
 } from '@aztec/stdlib/contract';
 import { Gas } from '@aztec/stdlib/gas';
-import { AbiDecodedSchema, type ApiSchemaFor, type ZodFor, optional, schemas } from '@aztec/stdlib/schemas';
+import { type ApiSchemaFor, type ZodFor, optional, schemas } from '@aztec/stdlib/schemas';
 import {
   Capsule,
   HashedValues,
@@ -29,12 +28,12 @@ import {
   TxReceipt,
   TxSimulationResult,
   UtilitySimulationResult,
-  inTxSchema,
 } from '@aztec/stdlib/tx';
-import type { ExecutionPayload, InTx } from '@aztec/stdlib/tx';
+import type { ExecutionPayload } from '@aztec/stdlib/tx';
 
 import { z } from 'zod';
 
+import { type Event, EventSchema, type PrivateEventFilter, PrivateEventFilterSchema } from '../api/events.js';
 import type {
   FeeEstimationOptions,
   GasSettingsOption,
@@ -133,47 +132,12 @@ export type BatchResults<T extends readonly BatchedMethod<keyof BatchableMethods
 };
 
 /**
- * Filter options when querying private events.
- */
-export type PrivateEventFilter = {
-  /** The address of the contract that emitted the events. */
-  contractAddress: AztecAddress;
-  /** Addresses of accounts that are in scope for this filter. */
-  scopes: AztecAddress[];
-  /** Transaction in which the events were emitted. */
-  txHash?: TxHash;
-  /** The block number from which to start fetching events (inclusive).
-   * Optional. If provided, it must be greater or equal than 1.
-   * Defaults to the initial L2 block number (INITIAL_L2_BLOCK_NUM).
-   * */
-  fromBlock?: BlockNumber;
-  /** The block number until which to fetch logs (not inclusive).
-   * Optional. If provided, it must be greater than fromBlock.
-   * Defaults to the latest known block to PXE + 1.
-   */
-  toBlock?: BlockNumber;
-};
-
-/**
- * An ABI decoded private event with associated metadata.
- */
-export type PrivateEvent<T> = {
-  /** The ABI decoded event */
-  event: T;
-  /** Metadata describing event context information such as tx and block */
-  metadata: InTx;
-};
-
-/**
  * The wallet interface.
  */
 export type Wallet = {
   getContractClassMetadata(id: Fr, includeArtifact?: boolean): Promise<ContractClassMetadata>;
   getContractMetadata(address: AztecAddress): Promise<ContractMetadata>;
-  getPrivateEvents<T>(
-    eventMetadata: EventMetadataDefinition,
-    eventFilter: PrivateEventFilter,
-  ): Promise<PrivateEvent<T>[]>;
+  getPrivateEvents<T>(eventMetadata: EventMetadataDefinition, eventFilter: PrivateEventFilter): Promise<Event<T>[]>;
   getChainInfo(): Promise<ChainInfo>;
   getTxReceipt(txHash: TxHash): Promise<TxReceipt>;
   registerSender(address: AztecAddress, alias?: string): Promise<AztecAddress>;
@@ -293,25 +257,6 @@ export const ContractClassMetadataSchema = z.object({
   artifact: z.union([ContractArtifactSchema, z.undefined()]),
 }) satisfies ZodFor<ContractClassMetadata>;
 
-export const EventMetadataDefinitionSchema = z.object({
-  eventSelector: schemas.EventSelector,
-  abiType: AbiTypeSchema,
-  fieldNames: z.array(z.string()),
-});
-
-export const PrivateEventSchema: ZodFor<PrivateEvent<AbiDecoded>> = z.object({
-  event: AbiDecodedSchema,
-  metadata: inTxSchema(),
-});
-
-export const PrivateEventFilterSchema = z.object({
-  contractAddress: schemas.AztecAddress,
-  scopes: z.array(schemas.AztecAddress),
-  txHash: optional(TxHash.schema),
-  fromBlock: optional(BlockNumberPositiveSchema),
-  toBlock: optional(BlockNumberPositiveSchema),
-});
-
 export const WalletSchema: ApiSchemaFor<Wallet> = {
   getChainInfo: z
     .function()
@@ -323,7 +268,7 @@ export const WalletSchema: ApiSchemaFor<Wallet> = {
   getPrivateEvents: z
     .function()
     .args(EventMetadataDefinitionSchema, PrivateEventFilterSchema)
-    .returns(z.array(PrivateEventSchema)),
+    .returns(z.array(EventSchema)),
   registerSender: z.function().args(schemas.AztecAddress, optional(z.string())).returns(schemas.AztecAddress),
   getAddressBook: z
     .function()
